+++ src/oddeven.h
@@ -22,8 +22,8 @@
  * @param array The array to sort
  * @param len The length of the array
  * @param size The size of a single element
- * @param cmp Comparison function that returns if an element is smaller than another
+ * @param cmp Element comparison function
  */
-void oddEvenSort(void** array, size_t len, size_t size, COMP_FUNC cmp);
+void oddEvenSort(void** array, int len, int size, COMP_FUNC cmp);
 
 #endif
